Add files via upload
This is my second attempt at this assignment, I'm not sure what happened the first time but I now did not submit it as a zip and I included a word document with an image of what it was initially supposed to look like because I know you said you have no idea what I did because it didn't work. So I included a screenshot. Unless I really didn't understand the assignment and this is wrong for the second time and you opened it the same the first time.